Cast one parameter to long when computing the area of the layout space
The value of the number which describes the layout area can quickly
exceed the integer range for large graphs. Therefore we use the long
data type to be on the safe side.

+++ src/main/java/org/rascat/gcl/layout/AbstractGraphCollectionLayout.java
@@ -18,8 +18,9 @@ public abstract class AbstractGraphCollectionLayout {
    *
    * @return the area
    */
-  public int area() {
-    return this.width * this.height;
+  public long area() {
+    // cast to long to prevent int overflow
+    return (long) this.width * this.height;
   }
 
   public int getHeight() {
